Refactored CreateEvent
